packages/pi-sensor/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
paho-mqtt>=2.0.0
RPi.GPIO>=0.7.1
adafruit-circuitpython-dht>=4.0.0
adafruit-circuitpython-ads1x15>=2.4.0
adafruit-circuitpython-tcs34725>=3.3.0
psutil>=5.9.0
requests>=2.31.0
